Cleaned up the navmesh and floors
The navmesh was kinda fuqed since their y levells were differrent and some of them were overlapping but it's mostly fine now. Also added an obstacle component for the cauldron to avoid it. You can also press M to spectate a monster currently just me and O for an overview